# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Geto API",
    "description": "Geto Database Management Tool API",
    "license": {
      "name": ""
    },
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/connections/{id}/completion": {
      "get": {
        "tags": [
          "crate::routes::meta"
        ],
        "operationId": "get_completion_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get auto-completion metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Connection not found"
          }
        }
      }
    },
...
    "/api/connections/{id}/databases": {
      "get": {
        "tags": [
          "crate::routes::meta"
        ],
        "operationId": "list_databases_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List databases in connection",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DatabaseInfo"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Connection not found"
          }
        }
      },
      "post": {
        "tags": [
          "crate::routes::databases"
        ],
        "operationId": "create_database_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatabaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Create new database",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDatabaseResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid database name"
          }
        }
      }
    },
    "/api/connections/{id}/databases/{name}": {
      "delete": {
        "tags": [
          "crate::routes::databases"
        ],
        "operationId": "drop_database_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Database name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Drop database",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DropDatabaseResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid database name"
          }
        }
      }
    },
...
    "/api/connections/{id}/schemas": {
      "get": {
        "tags": [
          "crate::routes::meta"
        ],
        "operationId": "list_schemas_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List schemas in connection",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Connection not found"
          }
        }
      }
    },
...
    "/api/connections/{id}/tables/{schema}/{table}": {
      "get": {
        "tags": [
          "crate::routes::meta"
        ],
        "operationId": "get_table_detail_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "schema",
            "in": "path",
            "description": "Schema name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "table",
            "in": "path",
            "description": "Table name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get table details (columns, indexes, constraints)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Table or connection not found"
          }
        }
      },
      "delete": {
        "tags": [
          "crate::routes::tables"
        ],
        "operationId": "drop_table_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "schema",
            "in": "path",
            "description": "Schema name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "table",
            "in": "path",
            "description": "Table name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Drop table",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DropTableResponse"
                }
              }
            }
          },
          "400": {
            "description": "Execution error"
          }
        }
      }
    },
...
    "/api/connections/{id}/tree": {
      "get": {
        "tags": [
          "crate::routes::meta"
        ],
        "operationId": "get_tree_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get database schema tree",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SchemaTree"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Connection not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ColumnInfo": {
        "type": "object",
        "required": [
          "name",
          "type",
          "notNull",
          "ordinal",
          "isPrimaryKey"
        ],
        "properties": {
          "default": {
            "type": [
              "string",
              "null"
            ]
          },
          "enumValues": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "isPrimaryKey": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "notNull": {
            "type": "boolean"
          },
          "ordinal": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "CompletionColumn": {
        "type": "object",
        "required": [
          "schema",
          "table",
          "name",
          "type"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "schema": {
            "type": "string"
          },
          "table": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
      "CompletionForeignKey": {
        "type": "object",
        "required": [
          "schema",
          "table",
          "columns",
          "refSchema",
          "refTable",
          "refColumns"
        ],
        "properties": {
          "columns": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "refColumns": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "refSchema": {
            "type": "string"
          },
          "refTable": {
            "type": "string"
          },
          "schema": {
            "type": "string"
          },
          "table": {
            "type": "string"
          }
        }
      },
      "CompletionFunction": {
        "type": "object",
        "required": [
          "schema",
          "name",
          "args",
          "returns",
          "kind"
        ],
        "properties": {
          "args": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "returns": {
            "type": "string"
          },
          "schema": {
            "type": "string"
          }
        }
      },
      "CompletionResponse": {
        "type": "object",
        "required": [
          "tables",
          "columns",
          "functions",
          "foreignKeys"
        ],
        "properties": {
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompletionColumn"
            }
          },
          "foreignKeys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompletionForeignKey"
            }
          },
          "functions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompletionFunction"
            }
          },
          "tables": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompletionTable"
            }
          }
        }
      },
      "CompletionTable": {
        "type": "object",
        "required": [
          "schema",
          "name",
          "type"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "schema": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/RelationType"
          }
        }
      },
...
      "ConstraintInfo": {
        "type": "object",
        "required": [
          "name",
          "type",
          "definition"
        ],
        "properties": {
          "definition": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
      "CreateDatabaseRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "CreateDatabaseResponse": {
        "type": "object",
        "required": [
          "created",
          "name"
        ],
        "properties": {
          "created": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "DatabaseInfo": {
        "type": "object",
        "required": [
          "name",
          "owner",
          "size"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "size": {
            "type": "string"
          }
        }
      },
...
      "DropDatabaseResponse": {
        "type": "object",
        "required": [
          "dropped"
        ],
        "properties": {
          "dropped": {
            "type": "boolean"
          }
        }
      },
      "DropTableResponse": {
        "type": "object",
        "required": [
          "dropped"
        ],
        "properties": {
          "dropped": {
            "type": "boolean"
          }
        }
      },
...
      "IndexInfo": {
        "type": "object",
        "required": [
          "name",
          "definition",
          "isUnique",
          "isPrimary"
        ],
        "properties": {
          "definition": {
            "type": "string"
          },
          "isPrimary": {
            "type": "boolean"
          },
          "isUnique": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "RelationEntry": {
        "type": "object",
        "required": [
          "name",
          "type"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/RelationType"
          }
        }
      },
      "RelationType": {
        "type": "string",
        "enum": [
          "table",
          "view",
          "matview"
        ]
      },
...
      "SchemaTree": {
        "type": "object",
        "required": [
          "schema",
          "relations"
        ],
        "properties": {
          "relations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RelationEntry"
            }
          },
          "schema": {
            "type": "string"
          }
        }
      },
...
      "TableDetailResponse": {
        "type": "object",
        "required": [
          "columns",
          "indexes",
          "constraints",
          "primaryKey",
          "foreignKeys"
        ],
        "properties": {
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ColumnInfo"
            }
          },
          "constraints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConstraintInfo"
            }
          },
          "foreignKeys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompletionForeignKey"
            }
          },
          "indexes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IndexInfo"
            }
          },
          "primaryKey": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}